+++ test/foundry/Invariants.t.sol
@@ -48,7 +48,7 @@ contract Invariants is Test {
     }
 
     function targetContracts() public view returns (address[] memory targetContracts_) {
-        require(_targetContracts.length != uint256(0), "NO_TARGET_CONTRACTS");
+        require(_targetContracts.length != uint(0), "NO_TARGET_CONTRACTS");
         return _targetContracts;
     }
 }
